requirements.txt
future==0.18.2
numpy==1.19.2
opencv-python==4.4.0.42
torch==1.6.0
grpcio
grpcio-tools
